added device image to flightcompanion
