+++ frontend/partner_devices/devices/BitsAndDroids/FlightCompanion.ts
@@ -6,4 +6,6 @@ export const FlightCompanion: PartnerDevice = {
   madeBy: "BitsAndDroids",
   bundle: FlightCompanionBundle,
   version: "1.0.0",
+  imageUrl:
+    "https://shop.bitsanddroids.com/cdn/shop/files/shop_overview-4.jpg?v=1718548830&width=713",
 };
